Add and fix compiler warnings

+++ include/tf-fs.h
@@ -2,6 +2,7 @@
 
 #include <stdio.h>
 #include <stdint.h>
+#include <stddef.h>
 
 #include "superblock.h"
 #include "bitmap.h"
@@ -18,10 +19,10 @@ typedef struct fs_t{
 inode_t *create_file(fs_t *fs, const char *name);
 
 /* Write data buffer to file. Will allocate block(s) if needed. */
-size_t write_file(fs_t *fs, inode_t *inode, uint8_t data[], size_t size, size_t offset);
+int write_file(fs_t *fs, inode_t *inode, uint8_t data[], size_t size, size_t offset);
 
 /* Reads file into out buffer. */
-size_t read_file(fs_t *fs, inode_t *inode, uint8_t out[], size_t size, size_t offset);
+int read_file(fs_t *fs, inode_t *inode, uint8_t out[], size_t size, size_t offset);
 
 /* Deletes file by freeing inode, bitmap, and blocks. */
 int delete_file(fs_t *fs, inode_t *inode);
